Extract references
