+++ __init__.py
@@ -6,3 +6,4 @@ from .features.configuration import *
 from .features.formatting import *
 from .features.definition import *
 from .features.hover import *
+from .features.references import *
